requirements.txt
fastapi
pydantic
uvicorn
redis
python-dotenv
sqlalchemy
psycopg2-binary
python-multipart
sqlmodel
tenacity
